fix: rename helpscout connector to helpscoutmailbox
